+++ app/src/main/java/pl/mattiahit/androidweatherk/database/daos/WeatherLocationDao.kt
@@ -1,13 +1,12 @@
-package pl.mattiahit.androidweatherk.models.daos
+package pl.mattiahit.androidweatherk.database.daos
 
-import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
-import pl.mattiahit.androidweatherk.models.WeatherLocation
+import pl.mattiahit.androidweatherk.database.entities.WeatherLocation
 
 @Dao
 interface WeatherLocationDao {
