+++ app/src/main/java/pl/mattiahit/androidweatherk/di/modules/LiveDataModule.kt
@@ -1,10 +1,9 @@
 package pl.mattiahit.androidweatherk.di.modules
 
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import pl.mattiahit.androidweatherk.WeatherApplication
-import pl.mattiahit.androidweatherk.livedata.LocationLiveData
+import pl.mattiahit.androidweatherk.home.domain.livedata.LocationLiveData
 import javax.inject.Singleton
 
 @Module
