+++ app/src/main/java/pl/mattiahit/androidweatherk/di/modules/RepositoryModule.kt
@@ -3,8 +3,8 @@ package pl.mattiahit.androidweatherk.di.modules
 import dagger.Module
 import dagger.Provides
 import pl.mattiahit.androidweatherk.WeatherApplication
-import pl.mattiahit.androidweatherk.repositories.LocationRepository
-import pl.mattiahit.androidweatherk.repositories.WeatherRepository
+import pl.mattiahit.androidweatherk.home.domain.repository.LocationRepository
+import pl.mattiahit.androidweatherk.home.domain.repository.WeatherRepository
 import javax.inject.Singleton
 
 @Module
@@ -12,13 +12,13 @@ class RepositoryModule(private val weatherApplication: WeatherApplication) {
 
     @Singleton
     @Provides
-    fun provideLocationRepository(): LocationRepository{
+    fun provideLocationRepository(): LocationRepository {
         return LocationRepository(weatherApplication)
     }
 
     @Singleton
     @Provides
-    fun provideWeatherRepository(): WeatherRepository{
+    fun provideWeatherRepository(): WeatherRepository {
         return WeatherRepository(weatherApplication)
     }
 }
