--- app/src/main/java/pl/mattiahit/androidweatherk/fragments/LocationWeatherFragment.kt
@@ -1,7 +1,0 @@
-package pl.mattiahit.androidweatherk.fragments
-
-
-class LocationWeatherFragment {
-
-
-}
