+++ app/src/main/java/pl/mattiahit/androidweatherk/home/domain/repository/LocationRepository.kt
@@ -1,11 +1,11 @@
-package pl.mattiahit.androidweatherk.repositories
+package pl.mattiahit.androidweatherk.home.domain.repository
 
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
 import pl.mattiahit.androidweatherk.WeatherApplication
 import pl.mattiahit.androidweatherk.database.AppDatabase
-import pl.mattiahit.androidweatherk.livedata.LocationLiveData
-import pl.mattiahit.androidweatherk.models.WeatherLocation
+import pl.mattiahit.androidweatherk.home.domain.livedata.LocationLiveData
+import pl.mattiahit.androidweatherk.database.entities.WeatherLocation
 import javax.inject.Inject
 
 class LocationRepository(val application: WeatherApplication) {
