+++ app/src/main/java/pl/mattiahit/androidweatherk/permission/PermissionListener.kt
@@ -1,4 +1,4 @@
-package pl.mattiahit.androidweatherk.utils
+package pl.mattiahit.androidweatherk.permission
 
 interface PermissionListener {
     fun isPermissionGranted(isGranted: Boolean)
